llm-based_language_learning_chatbot/backend/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
pydantic
openai
python-jose[cryptography]
passlib[bcrypt]
python-multipart
websockets
alembic